### 2020-09-17 #2683 @d-tear
I corrected some formatting errors in the bibliography and made some minor
edits to the acknowledgments and installation instructions
