### 2020-09-18 #2683 @jni
I don't have much bandwidth but progress is a bit stalled on my other reviews so, against my better judgment...

😂

@d-tear Thank you for the contribution and for the youtube videos, they are clear and informative! At a minimum you should add a citation to Fiji btw! 😉 Also, I'm not a reviewer (I'll work on finding some), but to get you a head start, I would say that (a) the summary should include one-line sentences for each of the three functions that the plugin provides, not just a motivation, (b) the statement of need should work a bit harder to differentiate the *specific* functions that were missing from Fiji — since one would assume that Fiji itself would have functionality for "the most common tasks that fellow wet-lab researchers encounter", and perhaps (c) mention a bit more clearly (perhaps a YouTube video) *how* SIAL would be integrated with QuPath and CP, since that is not clear to me at all.
