### 2020-09-18 #2683 @d-tear
@jni
<https://urldefense.com/v3/__https://github.com/jni__;!!LIr3w8kk_Xxm!4PQ8GtIFh6emehTCD4oiLzQW_yTE6-g9vK7daFwJqZnl60tJf2FyejyhCQim6g$>
thank
you for your recommended improvements to the paper! They are very
helpful (yep the absence of a citation to FIJI is a def an oversight). I
will work on incorporating these.
