### 2020-09-19 #2683 @d-tear
I incorporated those edits and included a new video showing how to import
FIJI/ImageJ rois into QuPath. This isn't a task that requires SIAL, but the
ROI Recorder program inside SIAL does make it easy to harvest FIJI/ImageJ
ROIs.

Please let me know if you have any more recommendations!
