### 2020-09-22 #2689 @whedon
[REVIEW]: SIAL: A simple image analysis library for wet-lab scientists
**Submitting author:** @d-tear (<a href="http://orcid.org/0000-0002-2281-6129">David Tyrpak</a>)
**Repository:** <a href="https://github.com/d-tear/SIAL" target ="_blank">https://github.com/d-tear/SIAL</a>
**Version:** v1.0.1
