### 2020-09-22 #2689 @jni
@bogovicj @haesleinhuepf thank you so much for agreeing to review this paper! You should have received a github invitation to join the joss-reviews repo which will allow you to tick items on your review checklist above. And, as mentioned in my email, the review process is also well-described in the [JOSS documentation](https://joss.readthedocs.io/en/latest/index.html). If you have any questions at all, feel free to ping me here!

@d-tear as above, feel free to ask me if you have any questions.
