### 2020-10-19 #2689 @jni
Hi everyone! Thank you for working on these reviews @bogovicj and @haesleinhuepf and thank you @d-tear for promptly addressing @haesleinhuepf's issues on the plugin repo! Can I get a status check from everyone on where things stand right now?
