### 2020-10-20 #2689 @bogovicj
@jni

There is a bit more left for me to do, though I expect to finish up this weekend at the latest.

A question - is this thread the right place to include suggestions for the Software paper? Or should those go elsewhere, say the project repo?
