### 2020-10-24 d-tear/SIAL#4 @haesleinhuepf
Community guidlines, links/references
Hi @d-tear ,

I'm just going through the checklist of the JOSS and see two minor issues I would like to ask you to take care of:
* Please put a short statement in readme of your SIAL repository telling users (and potentially developers) how they can reach you in case they need support or want to contribute to the project.
* Please put some links and references in your documentation. For example, you are mentioning Cell Profiler, QuPath and Fiji. Why not link to their respective websites?

Thanks!

Best,
Robert
