### 2020-10-24 d-tear/SIAL#5 @haesleinhuepf
Anonymization versus Pseudonymization
Hi @d-tear ,

I would like to ask you to think about clarifying some terminology in your software and in the documentation. In daily lab-life many people use the term "anonymization" but as later data analysis needs  association with the original names (e.g. patient names or file names), the correct term would be "pseudonymization" (see [wikipedia](https://en.wikipedia.org/wiki/Pseudonymization#:~:text=Pseudonymization%20is%20a%20data%20management,more%20artificial%20identifiers%2C%20or%20pseudonyms.)). Thus, I would ask you to at least clarify the difference in your documentation. I know people use these terms synonymously and that's why I think at least mentioning that there is a difference would be good to improve communication within the community.

Cheers,
Robert
