### 2020-10-24 #2689 @haesleinhuepf
Hi @jni, @bogovicj and @d-tear,

I'll just summarize my review in case I get distracted again.  ;-)

This project is a fantastic contribution to the ImageJ ecosystem. The three plugins for anonymisation (pseudonymisation?), phenotype-scoring and region-of-interest management are very useful, especiallly in the context of wet-lab imaging science. In many projects, automated image analysis is not at hand and programming experts not available. In such projects, manual anotation and phenotype scroring are state-of-the-art methods which are facilitated by the authors contribution in form a ImageJ/Fiji plugin. This eases the life of researchers and potentiallly reduces the likelyhood of manual errors.
I would like to thank the authors that they made these tools available to the public and open source. The documentation is very well done, partly in form of youtube videos very nicely demonstrating and explaining the usage of the plugins. Very well done. I have minor last issues: I would like to ask the authors to add a statement about [how to reach them in case someone needs support](https://github.com/d-tear/SIAL/issues/4), potentially some links to other scientific software mentioned in the github-readme documentation and adding a short [clarification of the difference between anonymization and pseudonymization](https://github.com/d-tear/SIAL/issues/5).

Otherwise, very well done! Please let me know if there is anything else I can do to push this project forward. I'm happy to help spreading the word btw.

Thanks!

Best,
Robert
