### 2020-10-24 d-tear/SIAL#6 @bogovicj
1.0.0 release, and avoid SNAPSHOT versions on update site
Hi @d-tear

When you're ready, please release version 1.0.0 and add it to your update site. And in the future,
avoid putting SNAPSHOT versions on your update site.

This is because it makes it difficult to determine the state of the source code that corresponds to the distributed jar. From Fiji [Development lifecycle](https://imagej.net/Development_Lifecycle) :

> Using a SNAPSHOT version is saying "no guarantees are made as to the reproducibility of this artifact." For this reason, to best facilitate reproducible science, SNAPSHOT versions of code are not provided to users (except potentially for testing).

Thanks!
John
