### 2020-10-24 #2689 @bogovicj
Hi @d-tear, @jni , @haesleinhuepf

Thanks David for your contribution!

Installation was easy and as described in the paper. The need for the software was well-motivated. For users, the tools are intuitive to use and quick to learn, especially given the video tutorials the authors provide.

As well, I quite like the logging mechanism that keeps track of what images users have already analyzed. That has the potential to be *super* useful for other developers.  I'd like to check it out soon (tomorrow?) - I may ask for some clarification about how a tool-builder might use it.

Minor comments regarding the paper:
I feel the description of the ROI recorder in the text is a little too brief to understand (from the text alone. That is, the text describes the goal of the roi recorder more than its functionality / use. Consider adding a couple short sentences describing what the tools do in action.

A more significant (but easy to address) issue has to do with what versions of the code you distribute via the fiji update site.  I'd ask that you [release a 1.0.0 version when you're ready and avoid distributing snapshot versions in the future.](https://github.com/d-tear/SIAL/issues/6)

Nicely done,
John
