### 2020-10-24 17d7035
Fixed version number and added comment to LogFile
I changed Pom version to 1.0.0
SNAPSHOTS shouldnt be distributed!
Also wrote a clarifying comment about private File infile in LogFile class. I dont
think this declaration is neccessary bc we are callign the parent FiilEditor constructor inside the LogFile constructor, and this parent constructor creates a private infile field. I want to do some tests though before I remove this line of code.
