### 2020-10-24 311f8dc
Update LogFile.java
clarified comment
