### 2020-10-24 6d8ac93
Update PhenoScoreKeeper.java
Number of phenotypes must be >= 2
