### 2020-10-24 ca8bf13
Update PhenoScoreKeeper.java
Code now ignores all other fields if a phenolog file is specified, and all variables are intialized from that chosen phenolog file
