### 2020-10-24 #2689 @d-tear
Hi @jni
<https://urldefense.com/v3/__https://github.com/jni__;!!LIr3w8kk_Xxm!_U7Nbabchvz-kBYleZmBizKicTUJQCQ1SmDnNak-HLRVjh5Jo0_mf2_mQLM5Fw$>
 , @haesleinhuepf
<https://urldefense.com/v3/__https://github.com/haesleinhuepf__;!!LIr3w8kk_Xxm!_U7Nbabchvz-kBYleZmBizKicTUJQCQ1SmDnNak-HLRVjh5Jo0_mf2_OP2FJRA$>,
and @bogovicj
<https://urldefense.com/v3/__https://github.com/bogovicj__;!!LIr3w8kk_Xxm!8_Sx2_UEdQ2GnjdLLT7MjlJ9A-o303ggvXymVKaKwk0DpfPqLHRH637LGEPh3w$>
,

I really appreciate your time and effort in improving the quality of the
paper and software!
This was my first foray into an open source software project and it's
really satisfying to see that this contribution would be beneficial to the
community.

I will incorporate your edits in the coming days, and address the issues in
the repository.
I'll also send an email out when I've addressed everything.

John, I'd be happy to discuss the logging mechanism code. The relevant
classes are the FileEditor and LogFile classes inside the repository.
I'd be happy to discuss the details with you or any other developers,
whether that's through email or zoom (note that I'm on the west coast PST).

Cheers,

David

On Sat, Oct 24, 2020 at 12:23 PM John Bogovic <notifications@github.com>
wrote:

> Hi @d-tear
> <https://urldefense.com/v3/__https://github.com/d-tear__;!!LIr3w8kk_Xxm!_U7Nbabchvz-kBYleZmBizKicTUJQCQ1SmDnNak-HLRVjh5Jo0_mf2-emoyPHQ$>,
> @jni
> <https://urldefense.com/v3/__https://github.com/jni__;!!LIr3w8kk_Xxm!_U7Nbabchvz-kBYleZmBizKicTUJQCQ1SmDnNak-HLRVjh5Jo0_mf2_mQLM5Fw$>
> , @haesleinhuepf
> <https://urldefense.com/v3/__https://github.com/haesleinhuepf__;!!LIr3w8kk_Xxm!_U7Nbabchvz-kBYleZmBizKicTUJQCQ1SmDnNak-HLRVjh5Jo0_mf2_OP2FJRA$>
>
> Thanks David for your contribution!
>
> Installation was easy and as described in the paper. The need for the
> software was well-motivated. For users, the tools are intuitive to use and
> quick to learn, especially given the video tutorials the authors provide.
>
> As well, I quite like the logging mechanism that keeps track of what
> images users have already analyzed. That has the potential to be *super*
> useful for other developers. I'd like to check it out soon (tomorrow?) - I
> may ask for some clarification about how a tool-builder might use it.
>
> Minor comments regarding the paper:
> I feel the description of the ROI recorder in the text is a little too
> brief to understand (from the text alone. That is, the text describes the
> goal of the roi recorder more than its functionality / use. Consider adding
> a couple short sentences describing what the tools do in action.
>
> A more significant (but easy to address) issue has to do with what
> versions of the code you distribute via the fiji update site. I'd ask that
> you release a 1.0.0 version when you're ready and avoid distributing
> snapshot versions in the future.
> <https://urldefense.com/v3/__https://github.com/d-tear/SIAL/issues/6__;!!LIr3w8kk_Xxm!_U7Nbabchvz-kBYleZmBizKicTUJQCQ1SmDnNak-HLRVjh5Jo0_mf2_V5yoPzQ$>
>
> Nicely done,
> John
>
