### 2020-10-25 01b0249
Update PhenoScoreKeeper.java
fixed typo
