### 2020-10-25 d-tear/SIAL#6 @d-tear
Hi @bogovicj ,

Thanks for brining this to my attention! It's a beginner mistake, my apologies. I removed the SNAPSHOT and have uploaded version 1.0.0 to my update site.
