+++ README.md
@@ -20,7 +20,7 @@ J.A. MacKay <sup>1,2,3</sup>
 
 In many biomedical research labs, image analysis tasks are relatively simple but labor intensive. For example, a typical workflow may require human intervention to outline regions of interest or score phenotypes from a collection of images. Such tasks could potentially be partially automated with a script or machine learning, but in our experience, many biomedical researchers do not obtain programming skills. Furthermore, image analysis is typically just one of the many experiments that a busy researcher will employ during the course of a project, and so devoting time to learning programming or bespoke image analysis is not prioritized. These factors create an unfortunate situation where, in addition to being labor intensive, image analysis workflows may become biased and exhibit limited reproducibility. 
 
-As primarily “wet-lab” scientists, we wanted to develop user friendly programs focused on the most common tasks that fellow wet-lab researchers encounter during image analysis. Because FIJI/ImageJ is routinely used by biomedical researchers the world over, we developed our programs in Java and packaged them together as a FIJI plugin. We name this plugin SIAL: A simple image analysis library. SIAL aids users in human-assisted image analysis by providing programs for image randomization, phenotype scoring, and ROI selection. Importantly, these programs automatically output useful metadata so that researchers can document their workflows. In addition, because wet-lab scientists are typically interrupted by multiple experiments, the programs keep track of which images in a directory have already been analyzed so that researchers can easily start and stop their workflows without hassle.
+As primarily “wet-lab” scientists, we wanted to develop user friendly programs focused on the most common tasks that fellow wet-lab researchers encounter during image analysis. Because <a href="https://fiji.sc">FIJI/ImageJ</a> is routinely used by biomedical researchers the world over, we developed our programs in Java and packaged them together as a FIJI plugin. We name this plugin SIAL: A simple image analysis library. SIAL aids users in human-assisted image analysis by providing programs for image randomization, phenotype scoring, and ROI selection. Importantly, these programs automatically output useful metadata so that researchers can document their workflows. In addition, because wet-lab scientists are typically interrupted by multiple experiments, the programs keep track of which images in a directory have already been analyzed so that researchers can easily start and stop their workflows without hassle.
 
 SIAL is easily installed via the ImageJ update website service, utilizes simple user interfaces, requires no programming experience, and requires no dependencies except FIJI. The individual programs inside SIAL can be easily integrated with workflows involving other FIJI plugins or with workflows employing another open source software, like Cell Profiler or QuPath. Because of its ease of use and focus on the most common wet-lab image analysis tasks, SIAL is routinely used by our group and our close collaborators to improve the efficieny and reproducibiltiy of our image analysis workflows. 
 
