### 2020-10-25 81966f3
Update README.md
added hyperlinks to Cell Profiler and QuPath
