### 2020-10-26 d-tear/SIAL#6 @bogovicj
Nice!

Can you please also tag the commit that corresponds to your 1.0.0 release?
You have one already for an older version (v0.1-Beta): https://github.com/d-tear/SIAL/tags

Thanks!
