### 2020-10-26 eaee186
Update README.md
added contact info and a new video link
