### 2020-10-26 d-tear/SIAL#6 @d-tear
Oh yeah that's def a good idea! I've tagged the commit that corresponds to the 1.0.0 release: https://github.com/d-tear/SIAL/tags
