### 2020-10-26 7a47433
Update paper.md
added clarification to what each of the three programs does.
