### 2020-10-26 3ecb253
Update paper.md
replaced randomization with the more accurate term "pseudonymization"
