+++ paper.md
@@ -32,7 +32,7 @@ bibliography: paper.bib
 
 In many biomedical research labs, image analysis tasks are relatively simple but labor intensive. For example, a typical workflow may require human intervention to outline regions of interest or score phenotypes from a collection of images. Such tasks could potentially be partially automated with a script or machine learning, but in our experience, many biomedical researchers do not obtain programming skills. Furthermore, image analysis is typically just one of the many experiments that a busy researcher will employ during the course of a project, and so devoting time to learning programming or bespoke image analysis is not prioritized. These factors create an unfortunate situation where, in addition to being labor intensive, image analysis workflows may become biased and exhibit limited reproducibility. 
 
-As primarily “wet-lab” scientists, we wanted to develop user friendly programs focused on some of the most common tasks that fellow wet-lab researchers encounter during image analysis. Because FIJI/ImageJ [@Schindelin:2012] is routinely used by biomedical researchers the world over, we developed our programs in Java and packaged them together as a FIJI plugin. We name this plugin SIAL: A simple image analysis library. SIAL aids users in human-assisted image analysis by providing programs for image randomization, phenotype scoring, and ROI harvesting. Each of the three programs inside SIAL is briefly summarized below:
+As primarily “wet-lab” scientists, we wanted to develop user friendly programs focused on some of the most common tasks that fellow wet-lab researchers encounter during image analysis. Because FIJI/ImageJ [@Schindelin:2012] is routinely used by biomedical researchers the world over, we developed our programs in Java and packaged them together as a FIJI plugin. We name this plugin SIAL: A simple image analysis library. SIAL aids users in human-assisted image analysis by providing programs for image pseudonymization, phenotype scoring, and ROI harvesting. Each of the three programs inside SIAL is briefly summarized below:
 
 File Randomizer: This program provides a convenient way for researchers to pseudonymize their imaging data, which is a critical step in any analysis involving phenotype scoring or ROI selection. The program copies images in a specified input directory to a specified oputput directory, and then randomly renames the copied images with a three digit number. It also outputs a key matching the original filenames to the renamed images and a details file which records time of analysis, as well as the specified input and output directory. 
 
