### 2020-10-27 bbcd202
Update paper.bib
corrected invalid DOI in FIJI reference
