+++ paper.bib
@@ -7,7 +7,7 @@
   	title = "{Fiji: an open-source platform for biological-image analysis}",
   	year = 2012,
     volume = 9,
-    doi = {https://doi.org/10.1038/nmeth.2019}
+    doi = {10.1038/nmeth.2019}
 }
 
 
