### 2020-10-27 12398cd
Update paper.md
fixed typos
