### 2020-10-27 7a5ff16
Update paper.md
more typos
