### 2020-10-27 82934ba
Update paper.md
added more correct term "pseudonymize"
