+++ paper.md
@@ -47,7 +47,7 @@ Both the PhenoScoreeKeeper and ROI Recorder programs track which images have alr
 
 Many biomedical researchers perform relatively simple image analysis tasks that often require human intervention. Notable examples include scoring phenotypes or harvesting ROIs from a directory of images. FIJI already provides many of the tools to carry out these analyses, but workflows are often compromised by user bias and limited reproducibility, as users may analyze data without proper blinding or neglect to save useful metadata. In addition, even these simple analyses can become labor intensive in the absence of a script which automates repetitive tasks while saving relevant output. 
 
-SIAL's goal is to provide simple GUI programs to help wet-lab researchers blind their data, score phenotypes, and harvest ROIs. In addition, because wet-lab scientists are typically interrupted by multiple experiments, the PhenoScoreKeeper and ROI Recorder programs keep track of which images in a directory have already been analyzed so that researchers can easily start and stop their workflows without hassle.
+SIAL's goal is to provide simple GUI programs to help wet-lab researchers blind and pseudonymize their data, score phenotypes, and harvest ROIs. In addition, because wet-lab scientists are typically interrupted by multiple experiments, the PhenoScoreKeeper and ROI Recorder programs keep track of which images in a directory have already been analyzed so that researchers can easily start and stop their workflows without hassle.
 
 
 SIAL is easily installed via the ImageJ update website service, utilizes simple user interfaces, requires no programming experience, and requires no dependencies except FIJI. The individual programs inside SIAL can be easily integrated with workflows involving other FIJI plugins or with workflows employing another open source software, like Cell Profiler [@McQuin:2018] or QuPath [@Bankhead:2017]. For example, users may employ the File Randomizer to blind their data before downstream analyses with these programs. Alternatively, within FIJI users may harvest ROIs using the ROI Recorder program and then analyze the ROIs using another plugin or import the ROIsets into QuPath for complementary analysis. 
