### 2020-10-27 a6de8dc
Update paper.md
shortened link
