### 2020-10-27 044b8f7
Update README.md
link to wikipedia article on pseudonymization, and restructured the readme to more closely mimic the paper.
