### 2020-10-27 6705554
Update paper.md
typo
