+++ paper.md
@@ -34,7 +34,7 @@ In many biomedical research labs, image analysis tasks are relatively simple but
 
 As primarily “wet-lab” scientists, we wanted to develop user friendly programs focused on some of the most common tasks that fellow wet-lab researchers encounter during image analysis. Because FIJI/ImageJ [@Schindelin:2012] is routinely used by biomedical researchers the world over, we developed our programs in Java and packaged them together as a FIJI plugin. We name this plugin SIAL: A simple image analysis library. SIAL aids users in human-assisted image analysis by providing programs for image pseudonymization, phenotype scoring, and ROI harvesting. Each of the three programs inside SIAL is briefly summarized below:
 
-File Randomizer: This program provides a convenient way for researchers to pseudonymize their imaging data, which is a critical step in any analysis involving phenotype scoring or ROI selection. The program copies images in a specified input directory to a specified oputput directory, and then randomly renames the copied images with a three digit number. It also outputs a key matching the original filenames to the renamed images and a details file which records time of analysis, as well as the specified input and output directory. 
+File Randomizer: This program provides a convenient way for researchers to pseudonymize their imaging data, which is a critical step in any analysis involving phenotype scoring or ROI selection. The program copies images in a specified input directory to a specified output directory, and then randomly renames the copied images with a three digit number. It also outputs a key matching the original filenames to the renamed images and a details file which records time of analysis, as well as the specified input and output directory. 
 
 PhenoScoreKeeper: This program helps speed up manual phenotype scoring of images by partially automating score collection. Users specify an input directory and output directory, and the program automatically opens up each image in the input directory and prompts the user to enter an integer score for the image. It saves the scores in a csv file located in the specified output directory.
 
