### 2020-10-27 21df53a
Update paper.md
typos
