### 2020-10-27 be36ec6
Update README.md
added additional clarification on pseudonymization vs anonymization
