### 2020-10-27 d-tear/SIAL#5 @d-tear
Hi @haesleinhuepf ,

Thanks for brining this to my attention. I've made efforts throughout the paper to use the more correct term of pseudonymization. In the repository readme, I also added a hyperlink to the wikipedia page for pseudonymization and added some additional sentences to differentiate pseudonymization vs anonymization
