### 2020-10-27 d-tear/SIAL#4 @d-tear
Hi @haesleinhuepf ,

These are great recommendations! In the repository readme I've added hyperlinks to FIJI, Cell Profiler, and QuPath (The paper contains references to the relevant publications). I also added a new section telling users/developers how they can contact me.
