### 2020-10-27 2f1ecfb
Update paper.md
formatting
