### 2020-10-27 6a08ed1
Update paper.md
corrected wordy sentence
