### 2020-10-27 #2689 @d-tear
Hi @jni , @bogovicj , and @haesleinhuepf ,

I believe I've addressed the open issues. If everything looks good, you can close them (or I can, either way).

(John, I went ahead closed the issue for the min number of phenotypes >=2. Let me know if that's OK. I also added some additional documentation for the ROI Recorder program.)

Please let me know if there's anything else to correct. I appreciate your valuable advice.
