+++ README.md
@@ -30,7 +30,7 @@ As primarily “wet-lab” scientists, we wanted to develop user friendly progra
 
 Both the PhenoScoreeKeeper and ROI Recorder programs track which images have already been analyzed so that users can stop and start their analyses without losing progress. Links to videos detailing the use of these programs are provided in the Installation and Tutorials section.
 
-SIAL is easily installed via the ImageJ update website service, utilizes simple user interfaces, requires no programming experience, and requires no dependencies except FIJI. The individual programs inside SIAL can be easily integrated with workflows involving other FIJI plugins or with workflows employing another open source software, like <a href="https://cellprofiler.org">Cell Profiler</a> or <a href="https://qupath.github.io">QuPath</a>. Because of its ease of use and focus on the most common wet-lab image analysis tasks, SIAL is routinely used by our group and our close collaborators to improve the efficiency and reproducibiltiy of our image analysis workflows. 
+SIAL is easily installed via the ImageJ update website service, utilizes simple user interfaces, requires no programming experience, and requires no dependencies except FIJI. The individual programs inside SIAL can be easily integrated with workflows involving other FIJI plugins or with workflows employing another open source software, like <a href="https://cellprofiler.org">Cell Profiler</a> or <a href="https://qupath.github.io">QuPath</a>. Because of its ease of use and focus on the most common wet-lab image analysis tasks, SIAL is routinely used by our group and our close collaborators to improve the efficiency and reproducibility of our image analysis workflows. 
 
 ### Author contributions:
 SIAL was developed by David R. Tyrpak at the MacKay lab in the School of Pharmacy of the University of Southern California. Yaocun Li and Siqi Lei tested the software and reported bugs.
