### 2020-11-14 #2689 @d-tear
Hi @jni
<https://urldefense.com/v3/__https://github.com/jni__;!!LIr3w8kk_Xxm!_U7Nbabchvz-kBYleZmBizKicTUJQCQ1SmDnNak-HLRVjh5Jo0_mf2_mQLM5Fw$>
 , @haesleinhuepf
<https://urldefense.com/v3/__https://github.com/haesleinhuepf__;!!LIr3w8kk_Xxm!_U7Nbabchvz-kBYleZmBizKicTUJQCQ1SmDnNak-HLRVjh5Jo0_mf2_OP2FJRA$>,
and @bogovicj
<https://urldefense.com/v3/__https://github.com/bogovicj__;!!LIr3w8kk_Xxm!8_Sx2_UEdQ2GnjdLLT7MjlJ9A-o303ggvXymVKaKwk0DpfPqLHRH637LGEPh3w$>
,

I hope you are doing well! I believe I've addressed all of the open issues
and concerns raised during the review process.
If there is anything else that needs to be addressed, please let me know.
Otherwise I will close the open issues in the repository.
@jni
<https://urldefense.com/v3/__https://github.com/jni__;!!LIr3w8kk_Xxm!_U7Nbabchvz-kBYleZmBizKicTUJQCQ1SmDnNak-HLRVjh5Jo0_mf2_mQLM5Fw$>
is
there anything else that needs to be addressed before the manuscript can be
published?

I appreciate all of your effort!

Best,

David
