### 2020-11-26 #2689 @jni
Hi everyone,

Apologies for the long silence, as it has been a very busy month for me. @bogovicj @haesleinhuepf thank you so much for your reviews so far. From the comments, it seems the paper is ready to be accepted, but I notice some unticked items in your reviewer checklists. Could you double check that everything is finished and confirm via comments that you are happy to accept the paper?

btw @d-tear it occurred to me this morning that for my very first paper, *at USC*, I scored confocal images ([Fig 4 of this paper](https://jvi.asm.org/content/80/6/2924/figures-only)), blinded by hand by my co-author Beth. I find it quite amusing to be reviewing this plugin 15 years later! 😂
