### 2020-11-27 #2689 @d-tear
@jni that is too funny. I actually have some friends that were in the Cannon lab. The world of science is so small!
