### 2020-12-08 205285d
Update README.md
Updated to the most recent (and correct) DOI for v1.0.0
