### 2020-12-09 #2689 @jni
Awesome! Thank you all!

@d-tear the next step is for you to generate a new tagged release + Zenodo archive, and report the release number and archive DOI here. Then we can set those versions accept the paper! 🎉
