### 2020-12-09 #2689 @d-tear
Hi @jni,

The new tagged release is v1.0.1 and the archive DOI is 10.5281/zenodo.4312250
