### 2020-12-09 #2689 @whedon
:wave: @openjournals/joss-eics, this paper is ready to be accepted and published.
