### 2020-12-14 #2689 @arfon
@bogovicj, @haesleinhuepf - many thanks for your reviews here and to @jni for editing this submission. JOSS relies upon the efforts of volunteers like yourselves, and we simply wouldn't be able to do it without you ✨
